sdk extraction > cleanup, add automatic conversion for UserProperty
