+++ src/main/java/com/hivemq/edge/modules/adapters/data/DataPoint.java
@@ -1,13 +1,13 @@
 package com.hivemq.edge.modules.adapters.data;
 
+import com.hivemq.extension.sdk.api.annotations.NotNull;
+
+/**
+ * Interface for data points gathered by a protocol adapter.
+ * Instances of it can be created via {@link com.hivemq.edge.modules.adapters.factories.DataPointFactory} accessible via {@link com.hivemq.edge.modules.adapters.factories.AdapterFactories} in the {@link com.hivemq.edge.modules.adapters.model.ProtocolAdapterInput}.
+ */
 public interface DataPoint {
-    Object getTagValue();
+    @NotNull Object getTagValue();
 
-    String getTagName();
-
-    @Override
-    boolean equals(Object o);
-
-    @Override
-    int hashCode();
+    @NotNull String getTagName();
 }
