--- src/main/java/com/hivemq/edge/modules/adapters/factories/ApiExceptionFactory.java
@@ -1,9 +1,0 @@
-package com.hivemq.edge.modules.adapters.factories;
-
-import com.hivemq.extension.sdk.api.annotations.NotNull;
-
-public interface ApiExceptionFactory {
-
-    @NotNull Exception create();
-
-}
