--- src/main/java/com/hivemq/edge/modules/api/adapters/ProtocolAdapterValidator.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2019-present HiveMQ GmbH
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.hivemq.edge.modules.api.adapters;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.hivemq.extension.sdk.api.annotations.NotNull;
-
-import java.util.List;
-
-/**
- * An object that handles the validation of the configuration bean, ensuring compliance with the requirement
- * schema
- *
- * @author Simon L Johnson
- */
-public interface ProtocolAdapterValidator {
-
-    @NotNull List<ProtocolAdapterValidationFailure> validateConfiguration(@NotNull ObjectMapper objectMapper, @NotNull Object config);
-
-}
