+++ src/main/java/com/hivemq/edge/modules/api/events/model/Event.java
@@ -1,11 +1,13 @@
 package com.hivemq.edge.modules.api.events.model;
 
-import com.hivemq.api.model.core.Payload;
 import com.hivemq.edge.model.Identifiable;
 import com.hivemq.edge.model.TypeIdentifier;
 import com.hivemq.extension.sdk.api.annotations.NotNull;
 import com.hivemq.extension.sdk.api.annotations.Nullable;
 
+/**
+ * Interface for events for the {@link com.hivemq.edge.modules.api.events.EventService}.
+ */
 public interface Event extends Identifiable {
     @NotNull SEVERITY getSeverity();
