--- src/main/java/com/hivemq/edge/modules/config/PollingConfiguration.java
@@ -1,8 +1,0 @@
-package com.hivemq.edge.modules.config;
-
-public interface PollingConfiguration {
-
-    long getPollingIntervalInMilliseconds();
-
-    int getMaxPollingErrorsBeforeRemoval();
-}
